fix visual
