+++ package.json
@@ -32,7 +32,7 @@
     "eslint": "^8",
     "eslint-config-next": "15.1.7",
     "postcss": "^8",
-    "prettier": "^3.4.2",
+    "prettier": "^3.5.2",
     "tailwindcss": "^3.4.1",
     "typescript": "^5"
   },
